# src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCF API",
    "summary": "API to calculate the predicted price for a vehicle with given features",
    "description": "\n## How to use this API\nlore ipsum dolor sit amet consectetur adipiscing elit\n\n## Available ENUMs\nlore ipsum dolor sit amet consectetur adipiscing elit\n",
    "contact": {
      "name": "Vadym Tatarinov",
      "url": "https://ford-parts.com.ru/",
      "email": "utikpuhlik@protonmail.ch"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local environment"
    },
    {
      "url": "https://api.dev.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Development environment"
    },
    {
      "url": "https://api.stage.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Staging environment"
    },
    {
      "url": "https://api.prod.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Production environment"
    }
  ],
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "operationId": "get_categories_categories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategorySchema"
                  },
                  "title": "Response Get Categories Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category",
        "operationId": "post_category_categories_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_category_categories_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/slug/{slug}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category by slug",
        "operationId": "get_category_by_slug_categories_slug__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CategorySchema"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Category By Slug Categories Slug  Slug  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category by id",
        "operationId": "get_category_by_id_categories__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CategorySchema"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Category By Id Categories  Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{category_id}": {
      "patch": {
        "tags": [
          "Categories"
        ],
        "summary": "Selective update category by id",
        "operationId": "patch_category_categories__category_id__patch",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_patch_category_categories__category_id__patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category by id",
        "operationId": "delete_category_categories__category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sub-categories": {
      "get": {
        "tags": [
          "Sub-Categories"
        ],
        "summary": "Get Sub Categories",
        "operationId": "get_sub_categories_sub_categories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "category_slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Category Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SubCategorySchema"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Sub Categories Sub Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sub-Categories"
        ],
        "summary": "Post Sub Category",
        "operationId": "post_sub_category_sub_categories_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_sub_category_sub_categories_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubCategorySchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sub-categories/{slug}": {
      "get": {
        "tags": [
          "Sub-Categories"
        ],
        "summary": "Get Sub Category By Slug",
        "operationId": "get_sub_category_by_slug_sub_categories__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SubCategorySchema"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Sub Category By Slug Sub Categories  Slug  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sub-categories/{sub_category_id}": {
      "put": {
        "tags": [
          "Sub-Categories"
        ],
        "summary": "Update a sub_category by id",
        "operationId": "put_sub_category_sub_categories__sub_category_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sub_category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sub Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_put_sub_category_sub_categories__sub_category_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubCategorySchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sub-Categories"
        ],
        "summary": "Delete sub_category by id",
        "operationId": "delete_sub_category_sub_categories__sub_category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sub_category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sub Category Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_patch_category_categories__category_id__patch": {
        "properties": {
          "category_payload": {
            "$ref": "#/components/schemas/CategoryPutSchema"
          },
          "image_blob": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Blob"
          }
        },
        "type": "object",
        "required": [
          "category_payload"
        ],
        "title": "Body_patch_category_categories__category_id__patch"
      },
      "Body_post_category_categories_post": {
        "properties": {
          "category_payload": {
            "$ref": "#/components/schemas/CategoryPostSchema"
          },
          "image_blob": {
            "type": "string",
            "format": "binary",
            "title": "Image Blob"
          }
        },
        "type": "object",
        "required": [
          "category_payload",
          "image_blob"
        ],
        "title": "Body_post_category_categories_post"
      },
...
      "Body_post_sub_category_sub_categories_post": {
        "properties": {
          "sub_category": {
            "$ref": "#/components/schemas/SubCategoryPostSchema"
          },
          "image_blob": {
            "type": "string",
            "format": "binary",
            "title": "Image Blob"
          }
        },
        "type": "object",
        "required": [
          "sub_category",
          "image_blob"
        ],
        "title": "Body_post_sub_category_sub_categories_post"
      },
...
      "Body_put_sub_category_sub_categories__sub_category_id__put": {
        "properties": {
          "sub_category": {
            "$ref": "#/components/schemas/SubCategoryPutSchema"
          },
          "image_blob": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Blob"
          }
        },
        "type": "object",
        "required": [
          "sub_category"
        ],
        "title": "Body_put_sub_category_sub_categories__sub_category_id__put"
      },
...
      "CategoryPostSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryPostSchema"
      },
      "CategoryPutSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryPutSchema"
      },
      "CategorySchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "examples": [
              "svechi-ford"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "slug"
        ],
        "title": "CategorySchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubCategoryPostSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи зажигания"
            ]
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "examples": [
              "2b3fb1a9-f13b-430f-a78e-94041fb0ed44"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id"
        ],
        "title": "SubCategoryPostSchema"
      },
      "SubCategoryPutSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи зажигания"
            ]
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "examples": [
              "2b3fb1a9-f13b-430f-a78e-94041fb0ed44"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id"
        ],
        "title": "SubCategoryPutSchema"
      },
      "SubCategorySchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Свечи зажигания"
            ]
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id",
            "examples": [
              "2b3fb1a9-f13b-430f-a78e-94041fb0ed44"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "https://storage.yandexcloud.net/tcf-images/default.svg"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "examples": [
              "svechi-zazhiganiia"
            ]
          },
          "category_slug": {
            "type": "string",
            "title": "Category Slug",
            "examples": [
              "svechi-ford"
            ]
          },
          "category_name": {
            "type": "string",
            "title": "Category Name",
            "examples": [
              "Свечи"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id",
          "id",
          "slug",
          "category_slug",
          "category_name"
        ],
        "title": "SubCategorySchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}